requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
click==6.7
cryptography==2.2.2
dominate==2.3.1
enum34==1.1.6
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
MySQL-python==1.2.5
paramiko==2.4.1
Pillow==5.2.0
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
python-dateutil==2.7.3
python-editor==1.0.3
regex==2018.2.21
six==1.11.0
SQLAlchemy==1.2.7
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
